requirements.txt
-i https://pypi.org/simple/

asyncio~=3.4.3
python-dotenv~=1.0.1
aiogram~=3.13.1
PyAutoGUI~=0.9.54